third_party/blink/web_tests/svg/dynamic-updates/SVGImageElement-dom-width-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'width' attribute of the SVGImageElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS imageElement.getAttribute('width') is "100"
PASS imageElement.getAttribute('width') is "200"
PASS successfullyParsed is true

TEST COMPLETE
